refactor(failover): consolidate QuickNode failover config into one module (#45306)
## **Description**

Follow-up to the network form shared failover work, addressing [mcmire's
review
comment](https://github.com/MetaMask/metamask-extension/pull/45230#discussion_r3730708680)
asking to consolidate the failover config and make it chain keyed.
Targets the feature branch, not main.

What moves into `shared/constants/network-failover.ts` (now the single
home for the QuickNode failover config):
- `QUICKNODE_ENDPOINT_URLS_BY_INFURA_NETWORK_NAME`
- `getFailoverUrlsForInfuraNetwork`
- `getIsQuicknodeEndpointUrl` (was in `network-utils.ts`)

`FEATURED_RPCS` switches from
`getFailoverUrlsForInfuraNetwork('bsc-mainnet')` to
`getFailoverUrlsForChainId(CHAIN_IDS.BSC)`, so you can tell at a glance
which chain each entry refers to.

### Breaking the circular import

Making `FEATURED_RPCS` (in `network.ts`) read from `network-failover.ts`
while `network-failover.ts` needs `CHAIN_IDS` from `network.ts` would
create a cycle that breaks the load time failover constant. To avoid it,
`CHAIN_IDS` moves into its own leaf module
`shared/constants/chain-ids.ts` that both import. `network.ts`
re-exports `CHAIN_IDS` and the moved failover symbols, so every existing
importer (including the frozen migrations 166 and 180) is untouched.

No behavior change. The circular deps check passes.

## **Changelog**

CHANGELOG entry: null

## **Related issues**

Fixes:

## **Manual testing steps**

1. This is a pure refactor with no behavior change; failover wiring and
the network form failover display work exactly as before.

## **Screenshots/Recordings**

### **Before**

### **After**

## **Pre-merge author checklist**

- [x] I've followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask
Extension Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've completed the PR template to the best of my ability
- [x] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [ ] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.

<!-- CURSOR_SUMMARY -->
---

> [!NOTE]
> **Medium Risk**
> Touches default RPC failover wiring across many chains and frozen
migrations; intended as behavior-preserving refactor but miswired env
getters could affect RPC fallback at runtime.
> 
> **Overview**
> QuickNode RPC failover configuration is **centralized** in
`shared/constants/network-failover.ts`, keyed by chain ID instead of
Infura network names. Removed from `network.ts` / `network-utils.ts`:
`QUICKNODE_ENDPOINT_URLS_BY_INFURA_NETWORK_NAME`,
`getFailoverUrlsForInfuraNetwork`, and the old
`getIsQuicknodeEndpointUrl` implementation.
> 
> **`CHAIN_IDS`** moves to a new leaf module
`shared/constants/chain-ids.ts` so `network-failover` and `network.ts`
can both use chain IDs without a circular import; `network.ts` still
re-exports `CHAIN_IDS` for existing importers.
> 
> Failover URLs are resolved via lazy env getters per chain
(`QUICKNODE_*_URL`). `FEATURED_RPCS` and related call sites now use
`getFailoverUrlsForChainId(CHAIN_IDS.*)`. Migrations **166** and **180**
inline `QUICKNODE_BASE_URL` instead of calling the removed helper. Tests
and `util.ts` import failover helpers from `network-failover`.
> 
> <sup>Reviewed by [Cursor Bugbot](https://cursor.com/bugbot) for commit
7efda51b42a052df13f844e540a31f201c1d701f. Bugbot is set up for automated
code reviews on this repo. Configure
[here](https://www.cursor.com/dashboard/bugbot).</sup>
<!-- /CURSOR_SUMMARY -->
