+++ shared/constants/network-failover.ts
@@ -1,82 +1,77 @@
 import { Hex } from '@metamask/utils';
-import {
-  CHAIN_IDS,
-  QUICKNODE_ENDPOINT_URLS_BY_INFURA_NETWORK_NAME,
-  getFailoverUrlsForInfuraNetwork,
-} from './network';
+import { CHAIN_IDS } from './chain-ids';
 
 /**
- * The chains that have a QuickNode failover, each paired with the Infura
- * network name used to resolve its failover URL. `as const` keeps every row a
- * fixed `[chainId, infuraNetworkName]` tuple, so callers can iterate and index
- * it without type assertions.
- *
- * This is the single source of truth shared between the NetworkController
- * initialization (which builds a chain ID keyed map) and the network form
- * (which looks up a single chain).
+ * The single source of truth for the QuickNode failover config: each chain that
+ * has a QuickNode failover, paired with a getter that reads its URL from the
+ * env lazily. Wrapping the env reads in getters means callers resolve the URLs
+ * when they run, so tests can just set `process.env` rather than reloading the
+ * module. `as const` keeps every row a fixed `[chainId, getUrl]` tuple so the
+ * helpers below can index it without type assertions.
  */
-const INFURA_NETWORK_NAME_BY_CHAIN_ID = [
-  [CHAIN_IDS.MAINNET, 'ethereum-mainnet'],
-  [CHAIN_IDS.LINEA_MAINNET, 'linea-mainnet'],
-  [CHAIN_IDS.ARBITRUM, 'arbitrum-mainnet'],
-  [CHAIN_IDS.AVALANCHE, 'avalanche-mainnet'],
-  [CHAIN_IDS.OPTIMISM, 'optimism-mainnet'],
-  [CHAIN_IDS.POLYGON, 'polygon-mainnet'],
-  [CHAIN_IDS.BASE, 'base-mainnet'],
-  [CHAIN_IDS.BSC, 'bsc-mainnet'],
-  [CHAIN_IDS.ZKSYNC_ERA, 'zksync-mainnet'],
-  [CHAIN_IDS.MEGAETH_MAINNET, 'megaeth-mainnet'],
-  [CHAIN_IDS.SEI, 'sei-mainnet'],
-  [CHAIN_IDS.MONAD, 'monad-mainnet'],
-  [CHAIN_IDS.HYPE, 'hyperevm-mainnet'],
-  [CHAIN_IDS.ARC, 'arc-mainnet'],
-  [CHAIN_IDS.ROBINHOOD_CHAIN, 'robinhood-mainnet'],
-] as const satisfies readonly (readonly [
-  Hex,
-  keyof typeof QUICKNODE_ENDPOINT_URLS_BY_INFURA_NETWORK_NAME,
-])[];
+const QUICKNODE_FAILOVER_URL_GETTERS_BY_CHAIN_ID = [
+  [CHAIN_IDS.MAINNET, () => process.env.QUICKNODE_MAINNET_URL],
+  [CHAIN_IDS.LINEA_MAINNET, () => process.env.QUICKNODE_LINEA_MAINNET_URL],
+  [CHAIN_IDS.ARBITRUM, () => process.env.QUICKNODE_ARBITRUM_URL],
+  [CHAIN_IDS.AVALANCHE, () => process.env.QUICKNODE_AVALANCHE_URL],
+  [CHAIN_IDS.OPTIMISM, () => process.env.QUICKNODE_OPTIMISM_URL],
+  [CHAIN_IDS.POLYGON, () => process.env.QUICKNODE_POLYGON_URL],
+  [CHAIN_IDS.BASE, () => process.env.QUICKNODE_BASE_URL],
+  [CHAIN_IDS.BSC, () => process.env.QUICKNODE_BSC_URL],
+  [CHAIN_IDS.ZKSYNC_ERA, () => process.env.QUICKNODE_ZKSYNC_URL],
+  [CHAIN_IDS.MEGAETH_MAINNET, () => process.env.QUICKNODE_MEGAETH_URL],
+  [CHAIN_IDS.SEI, () => process.env.QUICKNODE_SEI_URL],
+  [CHAIN_IDS.MONAD, () => process.env.QUICKNODE_MONAD_URL],
+  [CHAIN_IDS.HYPE, () => process.env.QUICKNODE_HYPEREVM_URL],
+  [CHAIN_IDS.ARC, () => process.env.QUICKNODE_ARC_URL],
+  [CHAIN_IDS.ROBINHOOD_CHAIN, () => process.env.QUICKNODE_ROBINHOOD_URL],
+] as const satisfies readonly (readonly [Hex, () => string | undefined])[];
 
 /**
- * The same mapping as an object for O(1) chain ID lookups. Built from the tuple
- * list so the two never drift. The URLs themselves are resolved lazily (they
- * depend on env), so only the chain ID to Infura network name lookup is cached
- * here.
- */
-const INFURA_NETWORK_NAME_BY_CHAIN_ID_LOOKUP: Partial<
-  Record<Hex, keyof typeof QUICKNODE_ENDPOINT_URLS_BY_INFURA_NETWORK_NAME>
-> = {};
-for (const [chainId, infuraNetworkName] of INFURA_NETWORK_NAME_BY_CHAIN_ID) {
-  INFURA_NETWORK_NAME_BY_CHAIN_ID_LOOKUP[chainId] = infuraNetworkName;
-}
-
-/**
- * Returns the QuickNode failover URLs for a chain, mirroring the map the
- * NetworkController is initialized with. Returns `undefined` for chains that
- * have no mapped failover so callers can fall back to an endpoint's own
- * `failoverUrls`.
+ * Returns the QuickNode failover URLs for a chain. Returns an empty array for a
+ * mapped chain whose env is unset, and `undefined` for a chain that has no
+ * mapped failover.
  *
  * @param chainId - The chain ID to look up.
  * @returns The failover URLs, or `undefined` if the chain has no mapped failover.
  */
 export function getFailoverUrlsForChainId(chainId: Hex): string[] | undefined {
-  const infuraNetworkName = INFURA_NETWORK_NAME_BY_CHAIN_ID_LOOKUP[chainId];
-  return infuraNetworkName
-    ? getFailoverUrlsForInfuraNetwork(infuraNetworkName)
-    : undefined;
+  const entry = QUICKNODE_FAILOVER_URL_GETTERS_BY_CHAIN_ID.find(
+    ([mappedChainId]) => mappedChainId === chainId,
+  );
+  if (!entry) {
+    return undefined;
+  }
+  const url = entry[1]();
+  return url ? [url] : [];
 }
 
 /**
- * Builds the chain ID to failover URLs map used to initialize the
+ * Builds the chain ID keyed QuickNode failover map used to initialize the
  * NetworkController. Chains whose QuickNode env is unset resolve to an empty
  * array.
  *
  * @returns A map of chain ID to failover URLs.
  */
 export function getFailoverUrlsByChainId(): Record<Hex, string[]> {
-  const failoverUrlsByChainId: Record<Hex, string[]> = {};
-  for (const [chainId, infuraNetworkName] of INFURA_NETWORK_NAME_BY_CHAIN_ID) {
-    failoverUrlsByChainId[chainId] =
-      getFailoverUrlsForInfuraNetwork(infuraNetworkName);
-  }
-  return failoverUrlsByChainId;
+  return QUICKNODE_FAILOVER_URL_GETTERS_BY_CHAIN_ID.reduce<
+    Record<Hex, string[]>
+  >((map, [chainId, getUrl]) => {
+    const url = getUrl();
+    map[chainId] = url ? [url] : [];
+    return map;
+  }, {});
+}
+
+/**
+ * Returns whether the given endpoint URL is one of the known QuickNode failover
+ * URLs.
+ *
+ * @param endpointUrl - The URL of the RPC endpoint.
+ * @returns True if the URL is a QuickNode URL, false otherwise.
+ */
+export function getIsQuicknodeEndpointUrl(endpointUrl: string): boolean {
+  return QUICKNODE_FAILOVER_URL_GETTERS_BY_CHAIN_ID.some(
+    ([, getUrl]) => getUrl() === endpointUrl,
+  );
 }
